homework/week1/arxiv_hunting/requirements.txt
requests
PyYAML
Click
